requirements.txt
numpy>=1.26
sounddevice>=0.4.6
soundfile>=0.12
tiktoken>=0.5.1
openai>=1.13.0
python-dotenv>=1.0.1
requests>=2.32.0

